Move update check into applyHandler and avoid deep clone.
The deep clone was because the update check was running in `commitHandler`, after the op is actually applied, and so needed a copy of the old doc. However, the `applyHandler` method is sufficient for our checking needs, and runs before the op is applied, avoiding the need for this deep copy. It also bring it into alignment with the create and delete checks.